+++ src/math/ray.cpp
@@ -8,6 +8,6 @@ const Vec3 &Ray::get_direction() const {
     return direction;
 }
 
-Vec3 Ray::at(float t) const {
+Vec3 Ray::at(double t) const {
     return origin + direction * t;
 }
